Move the Q.dir structure to the top level with root

+++ ZenPacks/zenoss/Import4/bin/cleanup_jobs.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#! /bin/bash
 ##############################################################################
 #
 # Copyright (C) Zenoss, Inc. 2014-2015, all rights reserved.
@@ -10,18 +10,29 @@
 
 export fail_records="/import4/perf.fail.records" # keep the failed status for top level UI
 
-# forcefully cleanup all the ongoing tasks/jobs
+# common block
+progdir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+source "$progdir/utils.sh"
 
+# save all the ongoing tasks/jobs at abort
 targets="
 tasks
 jobs
 tsdb
 "
 
-# remove the save_dir if exist
 save_dir="/import4/Q.save"
-rm -rf "$save_dir"
-rm -f "$fail_records"
+
+# save the last run info
+if [[ -d "$save_dir" ]] 
+then
+    chmod -R 777 "$save_dir"
+    rm -rf "$save_dir"
+fi
+
+mkdir -p  "$save_dir"
+chmod -R 777 "$save_dir"
+[ -f "$fail_records" ] && mv "$fail_records" "$save_dir"
 
 # the extra one for tsdb
 mkdir -p "/import4/Q.tsdb/.tmp"
@@ -32,7 +43,7 @@ chown -R zenoss:zenoss "/import4/Q.tsdb"
 for dname in $targets
 do
     # remove the residue directories
-    rm -rf "/import4/Q.$dname"
+    mv "/import4/Q.$dname" "$save_dir"
 
     # recreate the struct
     mkdir -p "/import4/Q.$dname/.done"
@@ -42,3 +53,5 @@ done
 
 # print out the dir structures
 find /import4 -type d -printf "%M %u %p"
+
+info_out "Performance directories cleaned up"
